requirement.txt
alembic==1.8.1
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==38.0.1
distlib==0.3.6
filelock==3.8.0
Flask==2.2.2
Flask-CKEditor==0.4.6
Flask-Login==0.6.2
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.1
greenlet==1.1.3
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.30
numpy==1.23.3
pi==0.1.2
platformdirs==2.5.2
protobuf==3.20.1
psycopg2==2.9.4
pycparser==2.21
PyMySQL==1.0.2
SQLAlchemy==1.4.41
virtualenv==20.16.5
Werkzeug==2.2.2
WTForms==3.0.1
